Filter out "tool options" actions in quick fix tests
When we use CodeInsightTestFixture#enableInspections in some test, it initializes inspection profile that is used across other tests. As a result, some annotators could register their actions during highlighting. Specifically, they register default actions to enable/disable and edit annotator settings. Such actions are not interesting when we check that no quick fixes are available. So we filter them out.